% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpcvis-package.r
\docType{package}
\name{hpcvis-package}
\alias{hpcvis-package}
\title{hpcvis}
\description{
Plot utilities for MPI codes and hardware counters.
}
\details{
\tabular{ll}{
   Package: \tab hpcvis\cr
   Type: \tab Package\cr
   License: \tab BSD 2-Clause\cr
}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}

