% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papi-plots.r
\name{papi-plot}
\alias{papi-plot}
\alias{plot.papi_cache}
\title{Plots PAPI objects}
\usage{
plot.papi_cache(x, ..., title, opnames, color = FALSE, show.opnames = TRUE,
  facet.by = "opnames")
}
\arguments{
\item{x}{PAPI object.}

\item{...}{Additional objects.}

\item{title}{Optional argument for adding a title to the plot.}

\item{opnames}{An optional argument to specify different names for the 
expressions/operations used to generate the profiler data.}

\item{color}{Logical; indicates whether groups should be colored differently
in the case of plotting multiple objects}

\item{show.opnames}{Logical; indicates whether or not expression/operation names should be
shown.}

\item{facet.by}{Choice to facet cache plots by the different expressions/operations
(\code{facet.by="opnames"}), or by the cache level (\code{facet.by="level"}).}
}
\description{
Plots PAPI objects
}
\examples{
\dontrun{
library(pbdPAPI)
x <- system.cache(rnorm(2e2))
y <- system.cache(rnorm(1e4))
z <- system.cache(rnorm(1e6))

library(scribe)
plot(x)
plot(x, show.opnames=FALSE)

plot(x, y, opnames=c("2e2", "1e4"), title="rnorm()")

plot(x, y, z, opnames=c("rnorm(2e2)", "rnorm(1e4)", "rnorm(1e6)"))

plot(x, y, z, opnames=c("2e2", "1e4"))
}

}

