% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scribe-package.r
\docType{package}
\name{scribe-package}
\alias{scribe-package}
\title{Benchmarking and Plot Utilities for Hardware Counters}
\description{
\tabular{ll}{
   Package: \tab scribe\cr
   Type: \tab Package\cr
   License: \tab BSD 2-Clause\cr
}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}

