% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachebench.r, R/cacheplot.r, R/papiplot.r
\name{papiplot.cachebench}
\alias{papiplot}
\alias{papiplot.cachebench}
\alias{papiplot.papi_cache}
\title{Cache Miss Benchmark Plotter}
\usage{
\method{papiplot}{cachebench}(x, ..., title, levels = 1:3,
  facet.by = "operation", label.angle = 0)

\method{papiplot}{papi_cache}(x, ..., title, opnames, color = FALSE,
  facet.by = "operation", bar.label = FALSE, label.angle = 0)

papiplot(x, ..., title, facet.by = "operation", label.angle = 0)
}
\arguments{
\item{x}{PAPI object.}

\item{...}{Additional objects.}

\item{title}{The label for the plot title. Should be a character string of
your choice, \code{NULL} for no label, or left blank for
the default plot label.  In the latter case, this is chosen
based on the input data.}

\item{levels}{Which cache levels to display.}

\item{facet.by}{Choice to facet cache plots by the different expressions/operations
(\code{facet.by="operation"}), or by the cache level (\code{facet.by="level"}).}

\item{label.angle}{The angle of x-axis labels.}

\item{opnames}{An optional argument to specify different names for the 
expressions/operations used to generate the profiler data.}

\item{color}{Logical; should different groups be colored?}

\item{bar.label}{Logical; should numeric values of heights of bars be shown?}

\item{title}{The label for the plot title. Should be a character string of
your choice, \code{NULL} for no label, or left blank for
the default plot label.  In the latter case, this is chosen
based on the input data.}
}
\value{
A ggplot2 object.
}
\description{
Cache Miss Benchmark Plotter

Plots Cacahe Misses

Create plots from PAPI performance counter data.
}
\details{
One may wonder why we do not simply overload \code{plot()}. In fact
that was the original incarnation of this package.  However,
overloading \code{plot()} leads to several issues.  Simply,
the basic plot command is *too* overloaded.  This makes it
difficult to find documentation, view function arguments, etc.
Using \code{papiplot()}, finding help is easier, and one can
also enjoy function argument autocompletion via the tab key.
}
\examples{
\dontrun{
library(pbdPAPI)
x <- cachebench(rnorm(1e4), rnorm(2e4), rnorm(3e4))

library(hpcvis)
papiplot(x)
papiplot(x, label.angle=15)
}

\dontrun{
library(pbdPAPI)
x <- system.cache(rnorm(1e4))
y <- system.cache(rnorm(4e4))
z <- system.cache(rnorm(8e4))

library(hpcvis)
papiplot(x)
papiplot(x, opnames=NULL)

opnames <- c("small", "medium", "large")
papiplot(x, y, z, opnames=opnames)
papiplot(x, y, z, opnames=opnames, color=TRUE, facet.by="level")
}

}

