% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof-plot.r
\name{profplot}
\alias{profplot}
\title{Plotting}
\usage{
profplot(x, which = 1L:4L, plot.type = "timing", title, bar.label = FALSE)
}
\arguments{
\item{x}{\code{prof} class object}

\item{which}{Vector consisting of a subset of the integers 1, 2, 3, 4.
Determines which plots will be produced.}

\item{plot.type}{for profiling mpiP including "timing", "stats1", "stats2",
"messages1", and "messages2".}

\item{title}{The label for the plot title. Should be a character string of
your choice, \code{NULL} for no label, or left blank for
the default plot label.  In the latter case, this is chosen
based on the input data.}

\item{bar.label}{logical; indicates whether or not numeric values of heights
of barplots should be included should .}
}
\description{
Plot methods for \code{prof} class objects.
}

